+++ tanks-game/oldbouncepower.h
@@ -52,8 +52,8 @@ public:
 
 class OldBounceBulletPower : public BulletPower {
 protected:
-	static const short maxBounces;
-	short bouncesLeft;
+	static const int maxBounces;
+	int bouncesLeft;
 
 public:
 	virtual void initialize(Bullet* parent) override;
